// src/sdk/database/orm/orm_form_submission.ts
export enum FormSubmissionFormType {
  Unspecified = 0,
  demo_request = 1,
  general_contact = 2,
}
// ...
export interface FormSubmissionModel {
  id: string;
  data_creator: string;
  data_updater: string;
  create_time: string;
  update_time: string;
  company_name: string;
  contact_person_name: string;
  email_address: string;
  phone_number?: string | null;
  message?: string | null;
  form_type: FormSubmissionFormType;
  submitted_at: string;
}